[FEAT] #81 주문 완료 누르면 주문 상태 화면으로 이동

+++ app/src/main/java/com/example/banchan/data/source/local/history/HistoryLocalDataSource.kt
@@ -19,7 +19,7 @@ class HistoryLocalDataSource @Inject constructor(
         Result.failure<Exception>(exception)
     }
 
-    override fun getHistoryWithItemsById(id: Int) = historyDao.getHistoryWithItemsById(id).map {
+    override fun getHistoryWithItemsById(id: Long) = historyDao.getHistoryWithItemsById(id).map {
         Result.success(it)
     }.catch { exception ->
         Result.failure<Exception>(exception)
